Refactor terminal command code into ui_terminal files
