+++ peer.c
@@ -1,4 +1,5 @@
 #include "peer.h"
+#include "ui_terminal.h"
 #include <stdarg.h>
 #include <signal.h>
 #include <unistd.h>
@@ -135,154 +136,6 @@ int add_peer(app_state_t *state, const char *ip, const char *username) {
     // We'r edone with the peer list so unlock
     pthread_mutex_unlock(&state->peers_mutex);
     return -1; // No space for new peer
-}
-
-void print_peers(app_state_t *state) {
-    // Lock the peers list whilst we use it
-    pthread_mutex_lock(&state->peers_mutex);
-    
-    time_t now = time(NULL);
-    int active_count = 0;
-    
-    printf("\n--- Active Peers ---\n");
-    for (int i = 0; i < MAX_PEERS; i++) {
-        if (state->peers[i].active) {
-            // Check if peer has timed out
-            if (now - state->peers[i].last_seen > PEER_TIMEOUT) {
-                state->peers[i].active = 0;
-                continue;
-            }
-            
-            printf("%d. %s@%s (last seen %ld seconds ago)\n", 
-                   ++active_count,
-                   state->peers[i].username,
-                   state->peers[i].ip, 
-                   now - state->peers[i].last_seen);
-        }
-    }
-    
-    if (active_count == 0) {
-        printf("No active peers found\n");
-    }
-    printf("------------------\n");
-    
-    pthread_mutex_unlock(&state->peers_mutex);
-}
-
-void *user_input_thread(void *arg) {
-    app_state_t *state = (app_state_t *)arg;
-    char input[BUFFER_SIZE];
-    char message[BUFFER_SIZE];
-    
-    printf("\nCommands:\n");
-    printf("  /list - List all active peers\n");
-    printf("  /send <peer_number> <message> - Send message to specific peer\n");
-    printf("  /broadcast <message> - Send message to all peers\n");
-    printf("  /quit - Exit the application\n\n");
-    
-    while (state->running) {
-        printf("> ");
-        fflush(stdout);
-        
-        if (fgets(input, BUFFER_SIZE, stdin) == NULL) {
-            if (state->running) {
-                log_message("Error reading input");
-            }
-            break;
-        }
-        
-        // Remove newline
-        input[strcspn(input, "\n")] = 0;
-        
-        if (strlen(input) == 0) {
-            continue;
-        }
-        
-        if (strcmp(input, "/list") == 0) {
-            print_peers(state);
-        } 
-        else if (strncmp(input, "/send ", 6) == 0) {
-            int peer_num;
-            char *msg_start = strchr(input + 6, ' ');
-            
-            if (msg_start == NULL) {
-                printf("Usage: /send <peer_number> <message>\n");
-                continue;
-            }
-            
-            *msg_start = '\0';
-            peer_num = atoi(input + 6);
-            msg_start++;
-            
-            pthread_mutex_lock(&state->peers_mutex);
-            
-            int count = 0;
-            int found = 0;
-            char target_ip[INET_ADDRSTRLEN];
-            
-            for (int i = 0; i < MAX_PEERS; i++) {
-                if (state->peers[i].active) {
-                    count++;
-                    if (count == peer_num) {
-                        strcpy(target_ip, state->peers[i].ip);
-                        found = 1;
-                        break;
-                    }
-                }
-            }
-            
-            pthread_mutex_unlock(&state->peers_mutex);
-            
-            if (found) {
-                if (send_message(target_ip, msg_start, MSG_TEXT) < 0) {
-                    log_message("Failed to send message to %s", target_ip);
-                } else {
-                    log_message("Message sent to %s", target_ip);
-                }
-            } else {
-                log_message("Invalid peer number");
-            }
-        }
-        else if (strncmp(input, "/broadcast ", 11) == 0) {
-            pthread_mutex_lock(&state->peers_mutex);
-            
-            for (int i = 0; i < MAX_PEERS; i++) {
-                if (state->peers[i].active) {
-                    if (send_message(state->peers[i].ip, input + 11, MSG_TEXT) < 0) {
-                        log_message("Failed to send message to %s", state->peers[i].ip);
-                    }
-                }
-            }
-            
-            pthread_mutex_unlock(&state->peers_mutex);
-            log_message("Broadcast message sent");
-        }
-        else if (strcmp(input, "/quit") == 0) {
-            // Send quit message to all peers before exiting
-            pthread_mutex_lock(&state->peers_mutex);
-            
-            for (int i = 0; i < MAX_PEERS; i++) {
-                if (state->peers[i].active) {
-                    // Send quit notification to each peer
-                    if (send_message(state->peers[i].ip, "", MSG_QUIT) < 0) {
-                        log_message("Failed to send quit notification to %s", state->peers[i].ip);
-                    }
-                }
-            }
-            
-            pthread_mutex_unlock(&state->peers_mutex);
-            log_message("Quit notifications sent to all peers");
-            
-            // Set running flag to false to exit the application
-            state->running = 0;
-            break;
-        }
-        else {
-            log_message("Unknown command. Type /list, /send, /broadcast, or /quit");
-        }
-    }
-    
-    return NULL;
 }
 
 int main(int argc, char *argv[]) {
